+++ lib/grape-swagger/endpoint.rb
@@ -271,7 +271,8 @@ module Grape
     end
 
     def build_root(route, reference, response_model, settings)
-      default_root = route.options[:is_array] ? response_model.downcase.pluralize : response_model.downcase
+      default_root = response_model.underscore
+      default_root = default_root.pluralize if route.options[:is_array]
       case route.settings.dig(:swagger, :root)
       when true
         { type: 'object', properties: { default_root => reference } }