+++ spec/lib/optional_object_spec.rb
@@ -39,7 +39,6 @@ describe GrapeSwagger::DocMethods::OptionalObject do
       let(:options) do
         { host: proc { |request| request.host =~ /^example/ ? '/api-example' : '/api' } }
       end
-      # rubocop:enable RegexpMatch
       specify do
         expect(subject.build(key, options, request)).to eql '/api-example'
       end